requirements.txt
# ── Infrastructure & Messaging ──────────────────────────────────
confluent-kafka==2.3.0

# ── Base de données ──────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# ── Data Science & ML ────────────────────────────────────────────
pandas==2.2.2
numpy==2.4.2
scikit-learn==1.4.2
xgboost==2.0.3

# ── API Backend ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.0

# ── Interface UI ─────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0

# ── Utilitaires ──────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.31.0
